backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
anthropic==0.20.0
openai==1.10.0
chromadb==0.4.22
sentence-transformers==2.3.1
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
pypdf2==3.0.1
pdfplumber==0.10.3
werkzeug==3.0.1
gunicorn==21.2.0
